+++ src/Stolons/Controllers/StolonsController.cs
@@ -29,29 +29,32 @@ namespace Stolons.Controllers
         }
 
         // GET: Stolons
-        [Authorize(Roles = Role_StolonAdmin + "," + Role_WedAdmin)]
-        public async Task<IActionResult> Index()
+        public IActionResult Index()
         {
-            return View(await _context.Stolons.ToListAsync());
+            if (!AuthorizedWebAdmin())
+                return Unauthorized();
+
+            return View(new StolonsViewModel(GetActiveAdherentStolon(), _context.Stolons.ToList()));
         }
 
         // GET: Stolons/Details/5
-        [Authorize(Roles = Role_StolonAdmin + "," + Role_WedAdmin)]
-        public async Task<IActionResult> Details(Guid? id)
+        public IActionResult Details(Guid? id)
         {
+            if (!Authorized(Role.Admin))
+                return Unauthorized();
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var stolon = await _context.Stolons
-                .SingleOrDefaultAsync(m => m.Id == id);
+            var stolon = _context.Stolons.FirstOrDefault(m => m.Id == id);
             if (stolon == null)
             {
                 return NotFound();
             }
+            return View(new StolonViewModel(stolon, GetCurrentAdherentSync().IsWebAdmin));
 
-            return View(stolon);
         }
 
         // GET: Stolons/Create
@@ -64,10 +67,12 @@ namespace Stolons.Controllers
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
-        [Authorize(Roles = Role_WedAdmin)]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(StolonViewModel vm, IFormFile uploadFile)
         {
+            if (!AuthorizedWebAdmin())
+                return Unauthorized();
+
             if (ModelState.IsValid)
             {
                 vm.Stolon.Id = Guid.NewGuid();
@@ -81,20 +86,22 @@ namespace Stolons.Controllers
 
 
         // GET: Stolons/Edit/5
-        [Authorize(Roles = Role_StolonAdmin + "," + Role_WedAdmin)]
-        public async Task<IActionResult> Edit(Guid? id, IFormFile uploadFile)
+        public IActionResult Edit(Guid? id)
         {
+            if (!Authorized(Role.Admin))
+                return Unauthorized();
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var stolon = await _context.Stolons.SingleOrDefaultAsync(m => m.Id == id);
+            var stolon = _context.Stolons.SingleOrDefault(x => x.Id == id);
             if (stolon == null)
             {
                 return NotFound();
             }
-            return View(new StolonViewModel(stolon));
+            return View(new StolonViewModel(stolon,GetCurrentAdherentSync().IsWebAdmin));
         }
 
         // POST: Stolons/Edit/5
@@ -102,7 +109,6 @@ namespace Stolons.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        [Authorize(Roles = Role_StolonAdmin + "," + Role_WedAdmin)]
         public async Task<IActionResult> Edit(Guid id, StolonViewModel vm, IFormFile uploadFile)
         {
             if (id != vm.Stolon.Id)
@@ -135,30 +141,32 @@ namespace Stolons.Controllers
         }
 
         // GET: Stolons/Delete/5
-        [Authorize(Roles = Role_WedAdmin)]
-        public async Task<IActionResult> Delete(Guid? id)
+        public IActionResult Delete(Guid? id)
         {
+            if (!AuthorizedWebAdmin())
+                return Unauthorized();
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var stolon = await _context.Stolons
-                .SingleOrDefaultAsync(m => m.Id == id);
+            var stolon = _context.Stolons.FirstOrDefault(m => m.Id == id);
             if (stolon == null)
             {
                 return NotFound();
             }
-
-            return View(stolon);
+            return View(new StolonViewModel( stolon, GetCurrentAdherentSync().IsWebAdmin));
         }
 
         // POST: Stolons/Delete/5
         [HttpPost, ActionName("Delete")]
-        [Authorize(Roles = Role_WedAdmin)]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
+            if (!AuthorizedWebAdmin())
+                return Unauthorized();
+
             var stolon = await _context.Stolons.SingleOrDefaultAsync(m => m.Id == id);
             _context.Stolons.Remove(stolon);
             await _context.SaveChangesAsync();
@@ -167,16 +175,18 @@ namespace Stolons.Controllers
 
 
 
-       
+
         [Route("Stolons/SwitchMode")]
-        [Authorize(Roles = Role_StolonAdmin + "," + Role_WedAdmin)]
         public IActionResult SwitchMode(Guid id)
         {
+            if (!Authorized(Role.Admin))
+                return Unauthorized();
+
             Stolon stolon = _context.Stolons.FirstOrDefault(x => x.Id == id);
             stolon.SimulationMode = stolon.GetMode() == Stolon.Modes.DeliveryAndStockUpdate ? Stolon.Modes.Order : Stolon.Modes.DeliveryAndStockUpdate;
             _context.Update(stolon);
             _context.SaveChanges();
-            return RedirectToAction("Details",new {id = id });
+            return RedirectToAction("Details", new { id = id });
         }
 
         private bool StolonExists(Guid id)
